feature - authentication for image repo

+++ uptane/repository/imagerepo.py
@@ -3,17 +3,23 @@ import os
 import json
 import zipfile
 import uptane.verify
+import uptane.crypto.sign
+import uptane.crypto.hash
 
 imagerepo = flask.Flask(__name__)
 imagerepo.config["UPLOAD_FOLDER"] = "public_repo"
 
-ROOT_METADATA_FILE_PATH:str
+ROOT_METADATA_FILE_PATH: str
+AUTH_PUB_ED25519_KEY: str
 
+
+# metadatafile verification lib functions
 def __snapshot_metadata_file_filter(value) -> bool:
     '''
     Keeps only snapshot metadata file in list
     '''
     return "snapshot" in value
+
 
 def __timestamp_metadata_file_filter(value) -> bool:
     '''
@@ -21,11 +27,13 @@ def __timestamp_metadata_file_filter(value) -> bool:
     '''
     return "snapshot" in value
 
-def get_snapshot_file_from_dirlist(file_list:list)->str:
+
+def get_snapshot_file_from_dirlist(file_list: list) -> str:
     '''
     gets the path of the snapshot file from the directory using uptane naming convention
     '''
-    snapshot_files:list[str] = list(filter(__snapshot_metadata_file_filter, file_list))
+    snapshot_files: list[str] = list(
+        filter(__snapshot_metadata_file_filter, file_list))
     if len(snapshot_files) == 0:
         raise Exception("no snaphot metadata file found")
 
@@ -34,20 +42,23 @@ def get_snapshot_file_from_dirlist(file_list:list)->str:
 
     return snapshot_files[0]
 
-def get_timestamp_file_from_dirlist(file_list:list)->str:
+
+def get_timestamp_file_from_dirlist(file_list: list) -> str:
     '''
     gets the path of the snapshot file from the directory using uptane naming convention
     '''
-    timestamp_files:list[str] = list(filter(__timestamp_metadata_file_filter, file_list))
+    timestamp_files: list[str] = list(
+        filter(__timestamp_metadata_file_filter, file_list))
     if len(timestamp_files) == 0:
         raise Exception("no timstamp metadata file found")
 
     if len(timestamp_files) > 1:
         raise Exception("more than one timestamp metadata file found")
-    
+
     return timestamp_files[0]
- 
-def remove_timestamp_file_from_repo(reponame:str)->None:
+
+
+def remove_timestamp_file_from_repo(reponame: str) -> None:
     '''
     removes latest timestamp file from repo for the new latest timestamp file
     '''
@@ -56,9 +67,10 @@ def remove_timestamp_file_from_repo(reponame:str)->None:
     except FileNotFoundError:
         pass
 
+
 # setting up different routes
 @imagerepo.route('/repo/<reponame>/<filename>/', methods=["GET", "POST"])
-def repo(reponame:str, filename:str):
+def repo(reponame: str, filename: str):
     global ROOT_METADATA_FILE_PATH
     try:
         if not os.path.exists('./public_repo/{}'.format(reponame)):
@@ -73,38 +85,62 @@ def repo(reponame:str, filename:str):
                     imagerepo.config["UPLOAD_FOLDER"],
                     '{}/{}'.format(reponame, filename))
         else:
+            # auth json will be of the form
+            # {"signed":{"hash":"hash of zip file", "bufsize":int}, "keyid":"KNaCaMgAlZnFePbHCuHgAgAuPt", "signature":"some_signature"}
+            auth_recv_dict = json.loads(flask.request.form["auth_json"])
+            print("auth recv: ", auth_recv_dict)
+            
+            # -----
+            # authentication
+            if auth_recv_dict["keyid"] != AUTH_PUB_ED25519_KEY:
+                return "", 401
+
+            if not uptane.crypto.sign.verify_sig_metadata(metadata=auth_recv_dict["signed"], \
+                hashf=uptane.crypto.hash.HashFunc.sha256, ktype=uptane.crypto.sign.KeyType.ed25519, \
+                pub_key=AUTH_PUB_ED25519_KEY, signature=auth_recv_dict["signature"]):
+                return "", 401
+
             file = flask.request.files["file"]
             # the file sent has to be a folder that has been zipped, not a files that have been zipped
 
             file.save('public_repo/{}/{}'.format(reponame, filename))
+            # now checking the file hash
+            if not uptane.crypto.hash.get_file_hash('public_repo/{}/{}'.format(reponame, filename), \
+                hashf=uptane.crypto.hash.HashFunc.sha256, bufsize=auth_recv_dict["signed"]["bufsize"]):
+                return "", 401
             
+            # --- 
             # unzipping the file
-            with zipfile.ZipFile('public_repo/{}/{}'.format(reponame, filename)) as zipO:
+            with zipfile.ZipFile('public_repo/{}/{}'.format(reponame,
+                                                            filename)) as zipO:
                 zipO.extractall('public_repo/{}'.format(reponame))
 
             # running the verification mechanism
             os.remove('public_repo/{}/{}'.format(reponame, filename))
-            
+
             # rewrite this ad-hoc code, as it forces to take .zip file name
-            verify_dir = 'public_repo/{}/{}'.format(reponame, filename.split('.')[0])           
+            verify_dir = 'public_repo/{}/{}'.format(reponame,
+                                                    filename.split('.')[0])
             recv_file_list = os.listdir(verify_dir)
             snapshot_file = get_snapshot_file_from_dirlist(recv_file_list)
             timestamp_file = get_timestamp_file_from_dirlist(recv_file_list)
-            
-            timestamp_file_path = '{}/{}'.format(verify_dir, timestamp_file) 
+
+            timestamp_file_path = '{}/{}'.format(verify_dir, timestamp_file)
             snapshot_file_path = '{}/{}'.format(verify_dir, snapshot_file)
-            
+
             # perform verification
             verifier = uptane.verify.Verification(root_metadata_file_path=ROOT_METADATA_FILE_PATH, \
                        timestamp_metadata_file_path=timestamp_file_path, \
                        snapshot_metadata_file_path=snapshot_file_path, \
                        targets_files_dir_path=verify_dir)
-            
+
             # move files to downloadable repo
             remove_timestamp_file_from_repo(reponame)
-            os.rename(timestamp_file_path, 'public_repo/{}/timestamp.toml'.format(reponame))
+            os.rename(timestamp_file_path,
+                      'public_repo/{}/timestamp.toml'.format(reponame))
             for file in os.listdir(verify_dir):
-                os.rename(verify_dir+'/'+file, 'public_repo/{}/{}'.format(reponame, file))
+                os.rename(verify_dir + '/' + file,
+                          'public_repo/{}/{}'.format(reponame, file))
 
             return '{"status":"success"}'
 
@@ -112,14 +148,13 @@ def repo(reponame:str, filename:str):
         return json.dumps({"error": {"type": str(e)}})
 
 
-def setup_server(root_metadata_file_path):
+def setup_server(root_metadata_file_path: str, authpubkey: str):
     global imagerepo
-    global ROOT_METADATA_FILE_PATH
+    global ROOT_METADATA_FILE_PATH, AUTH_PUB_ED25519_KEY
     # make python open up a specific directory in the filesystem
     if not os.path.exists('./public_repo'):
         os.mkdir("./public_repo")
-   
+
     ROOT_METADATA_FILE_PATH = root_metadata_file_path
+    AUTH_PUB_ED25519_KEY = authpubkey
     imagerepo.run(port=8080, debug=True)
-
-
